move risk calculator classes to folder
