+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessrisksandneeds/config/ONNXRuntimeConfig.kt
@@ -10,7 +10,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
 import org.springframework.core.io.Resource
-import uk.gov.justice.digital.hmpps.assessrisksandneeds.services.OnnxCalculatorServiceImpl
+import uk.gov.justice.digital.hmpps.assessrisksandneeds.services.riskCalculations.OnnxCalculatorServiceImpl
 
 @Profile("onnx-rsr")
 @Configuration
