+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessrisksandneeds/services/riskCalculations/OASysCalculatorServiceImpl.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.assessrisksandneeds.services
+package uk.gov.justice.digital.hmpps.assessrisksandneeds.services.riskCalculations
 
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Service
