+++ src/main/kotlin/uk/gov/justice/digital/hmpps/assessrisksandneeds/services/riskCalculations/SourceAnswers.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.assessrisksandneeds.services
+package uk.gov.justice.digital.hmpps.assessrisksandneeds.services.riskCalculations
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import com.fasterxml.jackson.annotation.JsonProperty
